Split common processes related GlobalSearch into a different file
